+++ Changelog.md
@@ -9,6 +9,7 @@ __New Features__
   - Updates defaults for `CompressorLockoutTemperature` and `BackupHeatingLockoutTemperature`.
   - Provides a warning if `BackupHeatingSwitchoverTemperature` or `BackupHeatingLockoutTemperature` are low and may cause unmet hours.
 - LightingGroups can now be specified using kWh/year annual consumption values as an alternative to fractions of different lighting types.
+- Allows building air leakage to be specified using CFMnatural or EffectiveLeakageArea.
 - Allows modeling one or more occupant vacancy periods (`VacancyPeriods` in the HPXML file).
 - ReportSimulationOutput measure:
   - Allows specifying the number of decimal places for timeseries output.